+++ capytaine/ui/cli.py
@@ -1,23 +1,17 @@
 #!/usr/bin/env python
 # coding: utf-8
 """Experimental command-line interface for Capytaine."""
-# Copyright (C) 2017-2019 Matthieu Ancellin
-# See LICENSE file at <https://github.com/mancellin/capytaine>
+# Copyright (C) 2017-2023 Matthieu Ancellin
+# See LICENSE file at <https://github.com/capytaine/capytaine>
 
 import argparse
-import logging
-import os
 
-from capytaine.io.legacy import import_cal_file, write_dataset_as_tecplot_files
-from capytaine.io.xarray import assemble_dataset
-from capytaine.bem.solver import BEMSolver
+import capytaine as cpt
+from capytaine.io.legacy import run_cal_file
 
-logging.basicConfig(level=logging.INFO,
-                    format="%(levelname)s:\t%(message)s")
+cpt.set_logging()
 
-LOG = logging.getLogger(__name__)
-
-parser = argparse.ArgumentParser(description="Command-line interface for the BEM solver Nemoh.")
+parser = argparse.ArgumentParser(description="Command-line interface for Capytaine taking Nemoh.cal files as input and returning Tecplots files.")
 parser.add_argument('paramfiles',
                     default=['./Nemoh.cal'],
                     nargs='*',
@@ -27,20 +21,7 @@ parser.add_argument('paramfiles',
 def main():
     args = parser.parse_args()
     for paramfile in args.paramfiles:
-        problems = import_cal_file(paramfile)
-        solver = BEMSolver()
-        results = [solver.solve(pb) for pb in problems]
-        data = assemble_dataset(results)
-        print(data)
-
-        results_directory = os.path.join(os.path.dirname(paramfile), 'results')
-        try:
-            os.mkdir(results_directory)
-        except FileExistsError:
-            LOG.warning("The 'results' directory already exists. You might be overwriting existing data.")
-
-        LOG.info("Write results in legacy tecplot format.")
-        write_dataset_as_tecplot_files(results_directory, data)
+        run_cal_file(paramfile)
 
 
 if __name__ == '__main__':